requirements.txt
streamlit
seleniumbase
webdriver-manager
lxml

